Move endpoints back into one mega file
